ResourceProgressBar -- show progress bar when gathering resources
* Display exp reward when killing a mob
* Smoother skillRoll probability increase with each new level (currently up to 64% at <skill level> 100)
* Update light when player destroys object by harvesting resources
* "fullscreen" launch flag

+++ MadSand-core/src/hitonoriol/madsand/entities/Entity.java
@@ -469,7 +469,7 @@ public abstract class Entity {
 
 	@JsonIgnore
 	public String getHealthState() {
-		float state = stats.hp / stats.mhp;
+		float state = (float) stats.hp / (float) stats.mhp;
 		if (state > HEALTH_75)
 			return HEALTH_STATE_FULL;
 		else if (state > HEALTH_50)
